can we use omni
recovery in this repo? It seems that I am educating myself more with recoveries & twrp, "if built correctly", likes to mount android 4.4 partitions better. Besides it is more user friendly & easy to use and understand. Unlike the boolean counter part, it works very well with android SDK. Furthermore we can use this repository for trial & error so that we may build a more stable rom that users can enjoy. Thanks for reading, this is really just a bunch of BS I wanted to write. Lol